Fix compile warnings with older gcc/g++ (#237)
* Fix compile warnings with older gcc/g++

With gcc/g++<10, there are some harmless warnings about redundant semicolon with macro expansion. Those warnings however prevent moose dbg mode (with -Werror) from building. So I will have to fix them here.

This will not be covered with github runners, because it's kinda difficult to pick an older compiler there. In the future, when we make the switch to CIVET for CI, we can add the test coverage back.

* Specify a range of supported cmake in pyproject.toml