Fix variable referencing bug (#258)
Wendy discovered a bug when a ScalarLinearInterpolation model is being used as a nonlinear parameter at multiple places. In that case, we issue a false-positive error saying the variable is referencing another variable multiple times.

This PR fixes that bug by relaxing that error checking to allow for repeated variable referencing iff the referee remains the same. The error message is also slightly improved to include more information.

The runner now supports subcommands to make it easier to diagnose models without a driver.